MSMF plugin

+++ modules/videoio/src/backend_plugin.cpp
@@ -465,7 +465,7 @@ public:
             unsigned n_params = (unsigned)(vint_params.size() / 2);
 
             if (CV_ERROR_OK == plugin_api->v2.Capture_open_with_params(
-                    filename.empty() ? 0 : filename.c_str(), camera, buffer.data(), buffer.size(),
+                    filename.empty() ? 0 : filename.c_str(), camera, buffer.data(), static_cast<uint32_t>(buffer.size()),
                     c_params, n_params, &capture))
             {
                 CV_Assert(capture);
